process metastable-resolved ADF15 files
