+++ aurora/atomic.py
@@ -174,244 +174,6 @@ def get_atom_data(imp, filetypes=['acd','scd'], filenames=[]):
     return atom_data
 
 
-def read_adf15(path, order=1, plot_lines=[], ax=None, Te_max = None, ne_max = None,
-               plot_log=False, plot_3d=False, pec_plot_min=None,  pec_plot_max = None):
-    """Read photon emissivity coefficients from an ADAS ADF15 file.
-
-    Returns a dictionary whose keys are the wavelengths of the lines in
-    angstroms. The value is an interp2d instance that will evaluate the PEC at
-    a desired density and temperature.
-
-    Args:
-        path : str
-            Path to adf15 file to read.
-        order : int, opt
-            Parameter to control the order of interpolation.
-        
-    Keyword Args:
-        plot_lines : list
-            List of lines whose PEC data should be displayed. Lines should be identified
-            by their wavelengths. The list of available wavelengths in a given file can be retrieved
-            by first running this function ones, checking dictionary keys, and then requesting a
-            plot of one (or more) of them.
-        ax : matplotlib axes instance
-            If not None, plot on this set of axes
-        plot_log : bool
-            When plotting, set a log scale
-        plot_3d : bool
-            Display PEC data as a 3D plot rather than a 2D one.
-        pec_plot_min : float
-            Minimum value of PEC to visualize in a plot
-        pec_plot_max : float
-            Maximum value of PEC to visualize in a plot
-        Te_max : float
-            Maximum Te value to plot when len(plot_lines)>1
-        ne_max : float
-            Maximum ne value to plot when len(plot_lines)>1
-
-    Returns:
-        pec_dict : dict
-            Dictionary containing interpolation functions for each of the available lines of the
-            indicated type (ionization or recombination). Each interpolation function takes as arguments
-            the log-10 of ne and Te.
-
-    Minimal Working Example (MWE)::
-
-        filename = 'pec96#h_pju#h0.dat' # for D Ly-alpha
-
-        # fetch file automatically, locally, from AURORA_ADAS_DIR, or directly from the web:
-        path = aurora.get_adas_file_loc(filename, filetype='adf15')  
-
-        # plot Lyman-alpha line at 1215.2 A. See available lines with pec_dict.keys() after calling without plot_lines argument
-        pec_dict = aurora.read_adf15(path, plot_lines=[1215.2])
-
-    Another example, this time also with charge exchange:::
-
-        filename = 'pec96#c_pju#c2.dat'
-        path = aurora.get_adas_file_loc(filename, filetype='adf15')
-        pec_dict = aurora.read_adf15(path, plot_lines=[361.7])
-
-    This function should work with PEC files produced via adas810 or adas218.
-
-    """
-    with open(path, 'r') as f:
-        lines = f.readlines()
-    cs = path.split('#')[-1].split('.dat')[0]
-
-    header = lines.pop(0)
-    # Get the expected number of lines by reading the header:
-    num_lines = int(header.split()[0])
-    pec_dict = {}
-
-    for i in range(0, num_lines):
-        # Get the wavelength, number of densities and number of temperatures
-        # from the first line of the entry:
-        l = lines.pop(0)
-        header = l.split()
-
-        # sometimes the wavelength and its units are not separated. Rough solution:
-        try:
-            #header = np.concatenate( [hh.split('A')[0] for hh in header])
-            header = [hh.split('A')[0] for hh in header]
-        except:
-            # lam and A are separated. Delete 'A' unit.
-            header = np.delete(header, 1)
-
-        try:
-            lam = float(header[0])
-        except ValueError:
-            # These lines appear to occur when lam has more digits than the
-            # allowed field width. For the moment, ignore these.
-            continue
-
-        try:
-            num_den = int(header[1])
-            num_temp = int(header[2])
-        except:
-            # 2nd element was empty -- annoyingly, this happens sometimes
-            num_den = int(header[2])
-            num_temp = int(header[3])
-
-        # Get the densities:
-        dens = []
-        while len(dens) < num_den:
-            dens += [float(v) for v in lines.pop(0).split()]
-        dens = scipy.asarray(dens)
-
-        # Get the temperatures:
-        temp = []
-        while len(temp) < num_temp:
-            temp += [float(v) for v in lines.pop(0).split()]
-        temp = scipy.asarray(temp)
-
-        # Get the PEC's:
-        PEC = []
-        while len(PEC) < num_den:
-            PEC.append([])
-            while len(PEC[-1]) < num_temp:
-                PEC[-1] += [float(v) for v in lines.pop(0).split()]
-        PEC = scipy.asarray(PEC)
-
-        # find what kind of rate we are dealing with
-        if 'recom' in l.lower():
-            rate_type = 'recom'
-        elif 'excit' in l.lower():
-            rate_type = 'excit'
-        elif 'chexc' in l.lower():
-            rate_type = 'chexc'
-            
-        # create dictionary with keys for each wavelength:
-        if lam not in pec_dict:
-            pec_dict[lam] = {}                
-
-        # add a key to the pec_dict[lam] dictionary for each type of rate: recom, excit or chexc
-        pec_dict[lam][rate_type] = scipy.interpolate.RectBivariateSpline(
-            np.log10(dens),
-            np.log10(temp),
-            PEC,
-            kx=order,
-            ky=order
-        )
-
-        if lam in plot_lines:
-
-            # use log spacing for ne values
-            if ne_max is not None:
-                # avoid overflow inside of np.logspace
-                ne_eval = 10** np.linspace(np.log10(dens.min()), np.log10(ne_max), 10)
-            else:
-                ne_eval = 10** np.linspace(np.log10(dens.min()), np.log10(dens.max()), 10)
-
-            # linear spacing for Te values
-            if Te_max is not None:
-                Te_eval = scipy.linspace(temp.min(), Te_max*1000, 100)
-            else:
-                Te_eval = scipy.linspace(temp.min(), temp.max(), 100)
-
-            NE, TE = scipy.meshgrid(ne_eval, Te_eval)
-
-            PEC_eval = pec_dict[lam][rate_type].ev(np.log10(NE), np.log10(TE))
-
-            # plot PEC rates
-            _plot_pec(dens,temp,ne_eval,Te_eval,PEC_eval,lam,cs,rate_type,
-                      ax, Te_max, ne_max, plot_log, plot_3d, pec_plot_min,  pec_plot_max)
-
-
-    return pec_dict
-
-
-
-def _plot_pec(dens, temp, ne_eval, Te_eval, PEC_eval,lam,cs,rate_type,
-             ax=None, Te_max = None, ne_max = None, plot_log=False, plot_3d=False, pec_plot_min=None,  pec_plot_max = None):
-    '''Private method to plot PEC data within :py:func:`~aurora.atomic.read_adf15` function.
-    '''
-
-    if ax is None:
-        f1 = plt.figure()
-        if plot_3d:
-            ax1 = f1.add_subplot(1,1,1, projection='3d')
-        else:
-            ax1 = f1.add_subplot(1,1,1)
-    else:
-        ax1 = ax
-
-    if plot_3d:
-        from mpl_toolkits.mplot3d import Axes3D
-
-        # linear scales (log doesn't work in matplotlib 3D)
-        # but can do all plotting on log quantities
-        if plot_log:
-            logNE, logTE = scipy.meshgrid(np.log10(ne_eval), np.log10(Te_eval))
-            ax1.plot_surface(logNE, logTE, PEC_eval, alpha=0.5)
-        else:
-            ax1.plot_surface(NE, TE, PEC_eval, alpha=0.5)
-
-        if plot_log:
-            dens = np.log10(dens); temp = np.log10(temp)
-
-        if Te_max is None and ne_max is None:
-            DENS, TEMP = scipy.meshgrid(dens, temp)
-            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC.T.ravel(), color='b')
-        elif Te_max is not None and ne_max is None:
-            Te_max_ind = np.argmin(np.abs(temp - Te_max*1000))
-            DENS, TEMP = scipy.meshgrid(dens, temp[:Te_max_ind+1])
-            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC[:,:Te_max_ind+1].T.ravel(), color='b')
-        elif Te_max is None and ne_max is not None:
-            ne_max_ind = np.argmin(np.abs(dens - ne_max))
-            DENS, TEMP = scipy.meshgrid(dens[:ne_max_ind+1], temp)
-            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC[:ne_max_ind+1,:].T.ravel(), color='b')
-        elif Te_max is not None and ne_max is not None:
-            Te_max_ind = np.argmin(np.abs(temp - Te_max*1000))
-            ne_max_ind = np.argmin(np.abs(dens - ne_max))
-            DENS, TEMP = scipy.meshgrid(dens[:ne_max_ind+1], temp[:Te_max_ind+1])
-            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC[:ne_max_ind+1,:Te_max_ind+1].T.ravel(), color='b')
-
-        if ax is None:
-            ax1.set_xlabel('$n_e$ [cm$^{-3}$]')
-            ax1.set_ylabel('$T_e$ [eV]')
-
-            ax1.set_zlabel('PEC')
-            ax1.set_title(cs + r' , $\lambda$ = '+str(lam) +' A, '+rate_type)
-            plt.tight_layout()
-
-    else:
-
-        # plot in 2D
-        labels = ['{:.0e}'.format(ne)+r' $cm^{-3}$' for ne in ne_eval]
-        for ine in np.arange(PEC_eval.shape[1]):
-            ax1.plot(Te_eval, PEC_eval[:,ine], label=labels[ine])
-        ax1.set_xlabel(r'$T_e$ [eV]')
-        ax1.set_ylabel('PEC')
-        ax1.set_yscale('log')
-        if pec_plot_min is not None:
-            ax1.set_ylim([pec_plot_min, ax1.get_ylim()[1]])
-        if pec_plot_max is not None:
-            ax1.set_ylim([ax1.get_ylim()[0], pec_plot_max ])
-
-        ax1.legend(loc='best').set_draggable(True)
-        ax1.set_title(cs + r' , $\lambda$ = '+str(lam) +' A, '+rate_type)
-
-
 
 
 def null_space(A):
