+++ aurora/radiation.py
@@ -1,6 +1,6 @@
 import os,sys
 import numpy as np
-from scipy.interpolate import interp1d
+from scipy.interpolate import RectBivariateSpline, interp1d
 from scipy.integrate import cumtrapz
 import matplotlib.pyplot as plt
 plt.ion()
@@ -571,6 +571,267 @@ def get_main_ion_dens(ne_cm3, ions, rhop_plot=None):
 
 
 
+def read_adf15(path, order=1, plot_lines=[], ax=None, Te_max = None, ne_max = None,
+               plot_log=False, plot_3d=False, pec_plot_min=None,  pec_plot_max = None):
+    """Read photon emissivity coefficients from an ADAS ADF15 file.
+
+    Returns a dictionary whose keys are the wavelengths of the lines in
+    angstroms. The value is an interp2d instance that will evaluate the PEC at
+    a desired density and temperature.
+
+    Args:
+        path : str
+            Path to adf15 file to read.
+        order : int, opt
+            Parameter to control the order of interpolation.
+        
+    Keyword Args:
+        plot_lines : list
+            List of lines whose PEC data should be displayed. Lines should be identified
+            by their wavelengths. The list of available wavelengths in a given file can be retrieved
+            by first running this function ones, checking dictionary keys, and then requesting a
+            plot of one (or more) of them.
+        ax : matplotlib axes instance
+            If not None, plot on this set of axes
+        plot_log : bool
+            When plotting, set a log scale
+        plot_3d : bool
+            Display PEC data as a 3D plot rather than a 2D one.
+        pec_plot_min : float
+            Minimum value of PEC to visualize in a plot
+        pec_plot_max : float
+            Maximum value of PEC to visualize in a plot
+        Te_max : float
+            Maximum Te value to plot when len(plot_lines)>1
+        ne_max : float
+            Maximum ne value to plot when len(plot_lines)>1
+
+    Returns:
+        pec_dict : dict
+            Dictionary containing interpolation functions for each of the available lines of the
+            indicated type (ionization or recombination). Each interpolation function takes as arguments
+            the log-10 of ne and Te.
+
+    Minimal Working Example (MWE)::
+
+        filename = 'pec96#h_pju#h0.dat' # for D Ly-alpha
+
+        # fetch file automatically, locally, from AURORA_ADAS_DIR, or directly from the web:
+        path = aurora.get_adas_file_loc(filename, filetype='adf15')  
+
+        # plot Lyman-alpha line at 1215.2 A. See available lines with pec_dict.keys() after calling without plot_lines argument
+        pec_dict = aurora.read_adf15(path, plot_lines=[1215.2])
+
+    Another example, this time also with charge exchange:::
+
+        filename = 'pec96#c_pju#c2.dat'
+        path = aurora.get_adas_file_loc(filename, filetype='adf15')
+        pec_dict = aurora.read_adf15(path, plot_lines=[361.7])
+
+    Metastable-resolved files will be automatically identified and parsed accordingly, e.g.
+
+         filename = 'pec96#he_pjr#he0.dat'
+         path = aurora.get_adas_file_loc(filename, filetype='adf15')
+         pec_dict = aurora.read_adf15(path, plot_lines=[584.4])
+
+    This function should work with PEC files produced via adas810 or adas218.
+
+    """
+    # find out whether file is metastable resolved
+    meta_resolved = path.split('#')[-2][-1]=='r'
+    if meta_resolved: print('Identified metastable-resolved PEC file')
+    
+    with open(path, 'r') as f:
+        lines = f.readlines()
+    cs = path.split('#')[-1].split('.dat')[0]
+
+    header = lines.pop(0)
+    # Get the expected number of lines by reading the header:
+    num_lines = int(header.split()[0])
+    pec_dict = {}
+
+    for i in range(0, num_lines):
+        # Get the wavelength, number of densities and number of temperatures
+        # from the first line of the entry:
+        l = lines.pop(0)
+        header = l.split()
+
+        # sometimes the wavelength and its units are not separated:
+        try:
+            header = [hh.split('A')[0] for hh in header]
+        except:
+            # lam and A are separated. Delete 'A' unit.
+            header = np.delete(header, 1)
+
+        lam = float(header[0])
+        # try:
+        #     lam = float(header[0])
+        # except ValueError:
+        #     # These lines appear to occur when lam has more digits than the
+        #     # allowed field width. For the moment, ignore these.
+        #     continue
+
+        if header[1]=='':
+            # 2nd element was empty -- annoyingly, this happens sometimes
+            num_den = int(header[2])
+            num_temp = int(header[3])
+        else:
+            num_den = int(header[1])
+            num_temp = int(header[2])            
+
+        if meta_resolved:
+            # index of metastable state
+            INDM = int(header[-3].split('/')[0].split('=')[-1])
+
+        # Get the densities:
+        dens = []
+        while len(dens) < num_den:
+            dens += [float(v) for v in lines.pop(0).split()]
+        dens = np.asarray(dens)
+
+        # Get the temperatures:
+        temp = []
+        while len(temp) < num_temp:
+            temp += [float(v) for v in lines.pop(0).split()]
+        temp = np.asarray(temp)
+
+        # Get the PEC's:
+        PEC = []
+        while len(PEC) < num_den:
+            PEC.append([])
+            while len(PEC[-1]) < num_temp:
+                PEC[-1] += [float(v) for v in lines.pop(0).split()]
+        PEC = np.asarray(PEC)
+
+        # find what kind of rate we are dealing with
+        if 'recom' in l.lower():
+            rate_type = 'recom'
+        elif 'excit' in l.lower():
+            rate_type = 'excit'
+        elif 'chexc' in l.lower():
+            rate_type = 'chexc'
+            
+        # create dictionary with keys for each wavelength:
+        if lam not in pec_dict:
+            pec_dict[lam] = {}                
+
+        # add a key to the pec_dict[lam] dictionary for each type of rate: recom, excit or chexc
+        pec_fun = RectBivariateSpline(
+            np.log10(dens),
+            np.log10(temp),
+            PEC,
+            kx=order,
+            ky=order
+        )
+
+        if meta_resolved:
+            if rate_type not in pec_dict[lam]:
+                pec_dict[lam][rate_type] = {}
+            pec_dict[lam][rate_type][f'meta{INDM}'] = pec_fun
+        else:
+            pec_dict[lam][rate_type] = pec_fun
+            
+        if lam in plot_lines:
+
+            # use log spacing for ne values
+            if ne_max is not None:
+                # avoid overflow inside of np.logspace
+                ne_eval = 10** np.linspace(np.log10(dens.min()), np.log10(ne_max), 10)
+            else:
+                ne_eval = 10** np.linspace(np.log10(dens.min()), np.log10(dens.max()), 10)
+
+            # linear spacing for Te values
+            if Te_max is not None:
+                Te_eval = np.linspace(temp.min(), Te_max*1000, 100)
+            else:
+                Te_eval = np.linspace(temp.min(), temp.max(), 100)
+
+            NE, TE = np.meshgrid(ne_eval, Te_eval)
+
+            PEC_eval = pec_fun.ev(np.log10(NE), np.log10(TE))
+
+            # plot PEC rates
+            _ax = _plot_pec(dens,temp,ne_eval,Te_eval,PEC_eval,lam,cs,rate_type,
+                            ax, Te_max, ne_max, plot_log, plot_3d, pec_plot_min,  pec_plot_max)
+
+            meta_str = ''
+            if meta_resolved: meta_str = f' , meta = {INDM}'
+            _ax.set_title(cs + r' , $\lambda$ = '+str(lam) +' A, '+rate_type+meta_str)
+            plt.tight_layout()
+
+    return pec_dict
+
+
+
+def _plot_pec(dens, temp, ne_eval, Te_eval, PEC_eval,lam,cs,rate_type,
+             ax=None, Te_max = None, ne_max = None, plot_log=False, plot_3d=False, pec_plot_min=None,  pec_plot_max = None):
+    '''Private method to plot PEC data within :py:func:`~aurora.atomic.read_adf15` function.
+    '''
+
+    if ax is None:
+        f1 = plt.figure()
+        if plot_3d:
+            ax1 = f1.add_subplot(1,1,1, projection='3d')
+        else:
+            ax1 = f1.add_subplot(1,1,1)
+    else:
+        ax1 = ax
+
+    if plot_3d:
+        from mpl_toolkits.mplot3d import Axes3D
+
+        # linear scales (log doesn't work in matplotlib 3D)
+        # but can do all plotting on log quantities
+        if plot_log:
+            logNE, logTE = np.meshgrid(np.log10(ne_eval), np.log10(Te_eval))
+            ax1.plot_surface(logNE, logTE, PEC_eval, alpha=0.5)
+        else:
+            ax1.plot_surface(NE, TE, PEC_eval, alpha=0.5)
+
+        if plot_log:
+            dens = np.log10(dens); temp = np.log10(temp)
+
+        if Te_max is None and ne_max is None:
+            DENS, TEMP = np.meshgrid(dens, temp)
+            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC.T.ravel(), color='b')
+        elif Te_max is not None and ne_max is None:
+            Te_max_ind = np.argmin(np.abs(temp - Te_max*1000))
+            DENS, TEMP = np.meshgrid(dens, temp[:Te_max_ind+1])
+            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC[:,:Te_max_ind+1].T.ravel(), color='b')
+        elif Te_max is None and ne_max is not None:
+            ne_max_ind = np.argmin(np.abs(dens - ne_max))
+            DENS, TEMP = np.meshgrid(dens[:ne_max_ind+1], temp)
+            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC[:ne_max_ind+1,:].T.ravel(), color='b')
+        elif Te_max is not None and ne_max is not None:
+            Te_max_ind = np.argmin(np.abs(temp - Te_max*1000))
+            ne_max_ind = np.argmin(np.abs(dens - ne_max))
+            DENS, TEMP = np.meshgrid(dens[:ne_max_ind+1], temp[:Te_max_ind+1])
+            ax1.scatter(DENS.ravel(), TEMP.ravel(), PEC[:ne_max_ind+1,:Te_max_ind+1].T.ravel(), color='b')
+
+        if ax is None:
+            ax1.set_xlabel('$n_e$ [cm$^{-3}$]')
+            ax1.set_ylabel('$T_e$ [eV]')
+            ax1.set_zlabel('PEC')
+
+    else:
+
+        # plot in 2D
+        labels = ['{:.0e}'.format(ne)+r' $cm^{-3}$' for ne in ne_eval]
+        for ine in np.arange(PEC_eval.shape[1]):
+            ax1.plot(Te_eval, PEC_eval[:,ine], label=labels[ine])
+        ax1.set_xlabel(r'$T_e$ [eV]')
+        ax1.set_ylabel('PEC')
+        ax1.set_yscale('log')
+        if pec_plot_min is not None:
+            ax1.set_ylim([pec_plot_min, ax1.get_ylim()[1]])
+        if pec_plot_max is not None:
+            ax1.set_ylim([ax1.get_ylim()[0], pec_plot_max ])
+
+        ax1.legend(loc='best').set_draggable(True)
+
+    return ax1
+
+
 def adf04_files():
     '''Collection of trust-worthy ADAS ADF04 files. 
     This function will be moved and expanded in ColRadPy in the near future. 
@@ -615,12 +876,12 @@ def adf04_files():
 
 
 
-def get_pec_prof(ion, cs, rhop, ne_cm3, Te_eV, lam_nm=1.8705, lam_width_nm=0.002, meta_idxs=[0],
+def get_colradpy_pec_prof(ion, cs, rhop, ne_cm3, Te_eV, lam_nm=1.8705, lam_width_nm=0.002, meta_idxs=[0],
                  adf04_repo = os.path.expanduser('~')+'/adf04_files/ca/ca_adf04_adas/',
                  pec_threshold=1e-20, phot2energy=True, plot=True):
     '''Compute radial profile for Photon Emissivity Coefficients (PEC) for lines within the chosen
     wavelength range using the ColRadPy package. This is an alternative to the option of using 
-    the :py:func:`~atomic.read_adf15` function to read PEC data from an ADAS ADF-15 file and 
+    the :py:func:`~aurora.radiation.read_adf15` function to read PEC data from an ADAS ADF-15 file and 
     interpolate results on ne,Te grids. 
 
     Args:
